Add lax and ops + fix tests related
