+++ jumpy/_custom_fns.py
@@ -18,10 +18,6 @@ except ImportError:
 
 F = TypeVar("F", bound=Callable)
 
-Carry = TypeVar("Carry")
-X = TypeVar("X")
-Y = TypeVar("Y")
-
 
 def array(object: Any, dtype=None) -> ndarray:
     """Creates an array given a list."""
@@ -34,30 +30,6 @@ def array(object: Any, dtype=None) -> ndarray:
         np = which_dtype(dtype)
 
     return np.array(object, dtype)
-
-
-def cond(
-    pred, true_fun: Callable[..., bool], false_fun: Callable[..., bool], *operands: Any
-):
-    """Conditionally apply true_fun or false_fun to operands."""
-    if is_jitted():
-        return jax.lax.cond(pred, true_fun, false_fun, *operands)
-    else:
-        if pred:
-            return true_fun(operands)
-        else:
-            return false_fun(operands)
-
-
-def fori_loop(lower: int, upper: int, body_fun: Callable[[X], X], init_val: X) -> X:
-    """Call body_fun over range from lower to upper, starting with init_val."""
-    if is_jitted():
-        return jax.lax.fori_loop(lower, upper, body_fun, init_val)
-    else:
-        val = init_val
-        for _ in range(lower, upper):
-            val = body_fun(val)
-        return val
 
 
 def index_update(x: ndarray, idx: ndarray, y: ndarray) -> ndarray:
@@ -80,47 +52,6 @@ def meshgrid(
         return onp.meshgrid(*xi, copy=copy, sparse=sparse, indexing=indexing)
 
 
-def scan(
-    f: Callable[[Carry, X], tuple[Carry, Y]],
-    init: Carry,
-    xs: X,
-    length: int | None = None,
-    reverse: bool = False,
-    unroll: int = 1,
-) -> tuple[Carry, Y]:
-    """Scan a function over leading array axes while carrying along state."""
-    if not is_jax_installed:
-        raise NotImplementedError("This function requires the jax module")
-
-    if is_jitted():
-        return jax.lax.scan(f, init, xs, length, reverse, unroll)
-    else:
-        xs_flat, xs_tree = jax.tree_util.tree_flatten(xs)
-        carry = init
-        ys = []
-        maybe_reversed = reversed if reverse else lambda x: x
-        for i in maybe_reversed(range(length)):
-            xs_slice = [x[i] for x in xs_flat]
-            carry, y = f(carry, jax.tree_util.tree_unflatten(xs_tree, xs_slice))
-            ys.append(y)
-        stacked_y = jax.tree_util.tree_map(lambda *y: onp.stack(y), *maybe_reversed(ys))
-        return carry, stacked_y
-
-
-def segment_sum(
-    data: ndarray, segment_ids: ndarray, num_segments: int | None = None
-) -> ndarray:
-    """Computes the sum within segments of an array."""
-    if which_np(data, segment_ids) is jnp:
-        s = jax.ops.segment_sum(data, segment_ids, num_segments)
-    else:
-        if num_segments is None:
-            num_segments = onp.amax(segment_ids) + 1
-        s = onp.zeros((num_segments,) + data.shape[1:])
-        onp.add.at(s, segment_ids, data)
-    return s
-
-
 def take(tree: Any, i: ndarray | Sequence[int] | int, axis: int = 0) -> Any:
     """Returns tree sliced by i."""
     if not is_jax_installed:
@@ -130,16 +61,6 @@ def take(tree: Any, i: ndarray | Sequence[int] | int, axis: int = 0) -> Any:
     if isinstance(i, (list, tuple)):
         i = np.array(i, dtype=int)
     return jax.tree_util.tree_map(lambda x: np.take(x, i, axis=axis, mode="clip"), tree)
-
-
-def top_k(operand: ndarray, k: int) -> tuple[ndarray, ndarray]:
-    """Returns top k values and their indices along the last axis of operand."""
-    if which_np(operand) is jnp:
-        return jax.lax.top_k(operand, k)
-    else:
-        top_ind = onp.argpartition(operand, -k)[-k:]
-        sorted_ind = top_ind[onp.argsort(-operand[top_ind])]
-        return operand[sorted_ind], sorted_ind
 
 
 def vmap(fun: F, include: Sequence[bool] | None = None) -> F:
@@ -183,16 +104,3 @@ def vmap(fun: F, include: Sequence[bool] | None = None) -> F:
         return jax.tree_util.tree_map(lambda *x: onp.stack(x), *rets)
 
     return _batched
-
-
-def while_loop(
-    cond_fun: Callable[[X], Any], body_fun: Callable[[X], X], init_val: X
-) -> X:
-    """Call body_fun while cond_fun is true, starting with init_val."""
-    if is_jitted():
-        return jax.lax.while_loop(cond_fun, body_fun, init_val)
-    else:
-        val = init_val
-        while cond_fun(val):
-            val = body_fun(val)
-        return val
